367: Reorganize models moving the commenting-responses to their own sub-directory
