+++ src/models/responses/commenting/index.ts
@@ -1,0 +1,4 @@
+export type {CommentModerationResponse} from "./CommentModerationResponse";
+export type {CommentReportResponse} from "./CommentReportResponse";
+export type {CommentResponse} from "./CommentResponse";
+export type {ReportResponse} from "./ReportResponse";
