+++ src/models/responses/index.ts
@@ -2,9 +2,6 @@ export type {AdminUserResponse} from "./AdminUserResponse";
 export type {AuditEntryResponse} from "./AuditEntryResponse";
 export type {BlockedDateResponse} from "./BlockedDateResponse";
 export type {CertificateResponse} from "./CertificateResponse";
-export type {CommentModerationResponse} from "./CommentModerationResponse";
-export type {CommentReportResponse} from "./CommentReportResponse";
-export type {CommentResponse} from "./CommentResponse";
 export type {DiveEventListItemResponse} from "./DiveEventListItemResponse";
 export type {DiveEventListResponse} from "./DiveEventListResponse";
 export type {DiveEventResponse} from "./DiveEventResponse";
@@ -24,11 +21,11 @@ export type {PaymentStatusResponse} from "./PaymentStatusResponse";
 export type {PortalConfigurationResponse} from "./PortalConfigurationResponse";
 export type {PortalConfigurationStatusResponse} from "./PortalConfigurationStatusResponse";
 export type {RegistrationResponse} from "./RegistrationResponse";
-export type {ReportResponse} from "./ReportResponse";
 export type {RolePermissionResponse} from "./RolePermissionResponse";
 export type {TermsResponse} from "./TermsResponse";
 export type {UserResponse} from "./UserResponse";
 
+export * from "./commenting";
 export * from "./downloads";
 export * from "./filetransfers";
 export * from "./stats";
